cleanup

+++ core/src/test/java/run/soeasy/framework/test/InMemoryPaginationsTest.java
@@ -23,7 +23,7 @@ public class InMemoryPaginationsTest {
 			Paginations<String> paginations = new Paginations<String>(Elements.of(list));
 			paginations.setPageSize(i);
 			Assert.assertTrue(
-					paginations.jumpToPage(paginations.getPages() + 1).getElements().count().longValue() == 0);
+					paginations.jumpToPage(paginations.getPages() + 1).getElements().count() == 0);
 			Assert.assertTrue(!paginations.jumpToPage(paginations.getPages()).hasNext());
 			Assert.assertArrayEquals(list.toArray(new String[0]),
 					paginations.all().getElements().toArray(String[]::new));
